Extract argv-array helpers to mcp/shared and fix SonarCloud findings
Codex pre-push review and SonarCloud both flagged the argv-helper
duplication between `mcp/ngfw/lib.js` and `mcp/ops/lib.js`. The
duplication was intentional in the first commit on this branch
(per-package npm boundaries) but it produced a 25% diff-duplication
and left two copies of an ADR-010-load-bearing contract free to drift.

Extract the helpers to `mcp/shared/aws-helpers.js` and re-export from
both `mcp/ngfw/lib.js` and `mcp/ops/lib.js`:

- `buildAwsArgv`, `awsExec`, `awsJson`, `awsText`,
  `buildSsmSendCommandArgs`, `REGION`, `getProfile` now live in one
  module that uses only `node:`-prefixed built-ins, so no
  cross-package npm dependency is needed.
- The shared `awsExec` carries the operation-labeled error format
  (`aws <service> <op>: <stderr>`) that codex flagged as needing
  alignment between the two packages — both now share it.
- `mcp/ops/lib.test.js` updated to expect the labeled error format.
- `mcp/ngfw/SECURITY.md` rewritten to describe the shared module
  rather than the previous duplication-acknowledgement narrative.
- ADR-010 evidence list now includes `mcp/shared/aws-helpers.js`.

Also fixes the SonarCloud code-smell findings on the new test files:

- `String.match()` calls replaced with `RegExp.exec()` (S6594).
- `arr[arr.length - 1]` replaced with `arr.at(-1)` (S7755).
- `void` operator removed; the variable was unused (S3735).
- IPV4_OCTET regex tightened to `\d` (S6353).

CHANGELOG entry under 3.95.17 expanded to cover the shared-module
extraction, the per-invocation key path, the EXIT-trap exit-code
preservation, the labeled errors, and the new shell-script
execution test (`mcp/ngfw/script-execution.test.js`).
