+++ docs/adr/index.yaml
@@ -268,6 +268,7 @@
     "exceptions": [],
     "enforcement": ["agent-policy", "pre-commit", "ci"],
     "evidence": [
+      "mcp/shared/aws-helpers.js",
       "mcp/ops/SECURITY.md",
       "mcp/ops/lib.js",
       "mcp/ops/index.js",
