+++ mcp/ops/lib.test.js
@@ -667,11 +667,15 @@ describe("awsExec", () => {
     );
   });
 
-  it("falls back to a generic message when stderr is empty and status is non-zero", () => {
+  it("falls back to a labeled generic message when stderr is empty and status is non-zero", () => {
+    // The shared awsExec helper labels every error with the AWS
+    // service+operation extracted from the argv (e.g.
+    // `aws s3 ls: exited with status 2`) so MCP handlers can
+    // localize which call failed.
     const runner = makeRecordingRunner({ status: 2 });
     assert.throws(
       () => awsExec("p", ["s3", "ls"], { runner }),
-      /aws exited with status 2/
+      /aws s3 ls: exited with status 2/
     );
   });
